LayoutTests/storage/indexeddb/transaction-storeNames-required-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test IndexedDB: transaction storeNames required arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
PASS db.transaction(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
